Shorten Union Tags
Makes it easier if someone ever has to type/look at them directly.

There is no need to make them more complex or Shaper specific.

Technically this could be a breaking change if someone somehow worked
with union tags manually. But I am not aware of anyone doing so.
And I highly encourage active userse to keep me in the loop if they are
doing interesting things.